Move tests doesn't depend on engine to kernel-api
