+++ kernel/kernel-api/src/test/scala/io/delta/kernel/internal/metadatadomain/TestJsonMetadataDomain.java
@@ -13,11 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.delta.kernel.defaults.utils;
+package io.delta.kernel.internal.metadatadomain;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import io.delta.kernel.internal.metadatadomain.JsonMetadataDomain;
 
 import java.util.Optional;
